refactor(pubsub): move helper functions to examples library
